+++ lib/strncasecmp.c
@@ -1,4 +1,5 @@
 #include <sys/types.h>
+#include <string.h>
 
 int  strncasecmp ( const char* s1, const char* s2, size_t len )
 {
@@ -13,7 +14,7 @@ int  strncasecmp ( const char* s1, const char* s2, size_t len )
         x1 = *s1++ - 'A'; if (x1 < 26u) x1 += 32;
         if ( x2 != x1 )
             break;
-        if ( x1 == -'A' )
+        if ( x1 == (unsigned int)-'A' )
             break;
     }
