+++ lib/strstr.c
@@ -2,8 +2,8 @@
 #include <string.h>
 
 char *strstr(const char *haystack, const char *needle) {
-  int nl=strlen(needle);
-  int hl=strlen(haystack);
+  size_t nl=strlen(needle);
+  size_t hl=strlen(haystack);
   int i;
   if (nl>hl) return 0;
   for (i=hl-nl+1; i; --i) {
